+++ src/types.client.ts
@@ -104,6 +104,11 @@ export interface WLEDClientSegment {
 	 */
 	repeat?:boolean
 
+	/**
+	 * If true the segment's contents will not be refreshed
+	 */
+	freeze?:boolean
+
 	/** Array that has up to 3 color arrays as elements, the primary, secondary (background) and tertiary colors of the segment. Each color is an array of 3 or 4 bytes, which represent an RGB(W) color. */
 	colors?:(RGBColor|RGBWColor)[]